SALTO-2309: move test helpers from salesforce index (#3462)
